# api/pb/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/broadcast": {
      "post": {
        "operationId": "Broadcast",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbSimpleMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbBroadcastMessage"
            }
          }
        ],
        "tags": [
          "SpaceMeshService"
        ]
      }
    },
    "/v1/example/echo": {
      "post": {
        "operationId": "Echo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbSimpleMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbSimpleMessage"
            }
          }
        ],
        "tags": [
          "SpaceMeshService"
        ]
      }
    },
    "/v1/register": {
      "post": {
        "operationId": "RegisterProtocol",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbSimpleMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbProtocol"
            }
          }
        ],
        "tags": [
          "SpaceMeshService"
        ]
      }
    },
    "/v1/send": {
      "post": {
        "operationId": "SendMessage",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbSimpleMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbInMessage"
            }
          }
        ],
        "tags": [
          "SpaceMeshService"
        ]
      }
    }
  },
  "definitions": {
    "pbBroadcastMessage": {
      "type": "object",
      "properties": {
        "protocolName": {
          "type": "string"
        },
        "payload": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "pbInMessage": {
      "type": "object",
      "properties": {
        "nodeID": {
          "type": "string"
        },
        "protocolName": {
          "type": "string"
        },
        "payload": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "pbProtocol": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "port": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "pbSimpleMessage": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    }
  }
}